knative/Video/vi-extract/requirements.txt
Flask==2.2.3
minio==7.1.13
parliament-functions==0.1.0
